Feat: Envelope Listener
This will check if the envelope (ItemStack Paper) was sealed or not.  If the envelope was not sealed then it will open up a custom inventory with seal button in slot 8. Once the player selects the seal button the item will apply a default title of "Sealed" and place the envelope into the players hand. Then if the player right clicks the envelope again the player will be prompt a message using the "Conversation API" asking to set a custom title. Once title is set the item will remove the lore stating to set a title and set a custom data "1337" to ensure that the player cannot unseal or open its inventory again.