Feat: Mailbox listener
Upon placing a Player head it will create a custom inventory named "Players Name + Mailbox" and set the players UUID as a identifier. However, for MVP made it so that you must have a "Dirt Block" in your off hand for the player head to turn into a mailbox.

When the mailbox is clicked (right clicked) the mailbox inventory will open unless it is empty will output a msg stating it is empty. When the mailbox inventory opens and the player clicks the envelope, the items contained inside the envelope inventory will automatically be placed inside the players inventory. The players inventory will be checked to ensure it is not full and if it is full then it will check for any stackable items. If the envelope becomes empty it will be removed from the players mailbox.

 If the player does not own the mailbox the players inventory will be checked to ensure they are holding a envelope(paper) and that it has a custom data of "1337" and if it does then the envelope will automatically be placed into that players mailbox.

When the player joins and they have an envelope inside their mailbox they will receive a message.