Feat: MVC Abstraction per request
Removed all implementation and only added a MVC Abstraction approach for simplicity. This was a request per GuoKao. The following framework will allow for creating a inventory and a holder. The holder and inventory will be rendered for easy abstraction. When creating the inventory, the identifier will be put into a HashMap for getting letter. To add items into the inventory can be easily done with the Container class which handles items being placed along with its position, which will also render the inventory and holder. Removed all other classes.